Revert function return type for ABI consistency
We want to keep psmove_get_temperature_in_celsius() compatible to
previous versions, so the function return value is changed back from
int to the original float.

+++ include/psmove.h
@@ -634,7 +634,7 @@ ADDCALL psmove_get_temperature(PSMove *move);
  *
  * \return The temperature sensor reading in degree Celcius
  **/
-ADDAPI int
+ADDAPI float
 ADDCALL psmove_get_temperature_in_celsius(PSMove *move);
 
 /**
